producer/app/requirements.txt
tweepy
kafka-python
configparser
joblib